Remove InfluxDB v1 metrics middleware
