+++ docs/content/reference/static-configuration/file.yaml
@@ -298,20 +298,6 @@ metrics:
     addRoutersLabels: true
     addServicesLabels: true
     prefix: foobar
-  influxDB:
-    address: foobar
-    protocol: foobar
-    pushInterval: 42s
-    database: foobar
-    retentionPolicy: foobar
-    username: foobar
-    password: foobar
-    addEntryPointsLabels: true
-    addRoutersLabels: true
-    addServicesLabels: true
-    additionalLabels:
-      name0: foobar
-      name1: foobar
   influxDB2:
     address: foobar
     token: foobar
